ref: split tab-content into components
